+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Translation of `pandas.Index` types.
 - The `translation.testing` module.
 - Experimental and hacky implementation of translation for nested sequences
 - Entry point `rics-perf` for multivariate performance testing, taking candidates from `./candidates.py`